fix(GenUI): display action buttons for nested cards and tables
- Move action button rendering into ComponentRendererInner to support nested components
- Action buttons (Download PNG for cards, Copy/Download CSV for tables) now render regardless of nesting level
- Extract ActionButtonsContainer to reduce code duplication
- Wrap GenUIActionLink with memo() for performance optimization
- Use useCallback for all handler functions with proper dependency tracking
- Remove duplicate action button logic from non-block-level component return path
- Wrap components in React.Fragment for proper key handling

This fixes the issue where action buttons were only visible for top-level cards and tables, not when nested inside other components like STACK.
